chromecast-wrapper: Check if the sub file exists.

+++ chromecast-wrapper.sh
@@ -131,6 +131,8 @@ else
 
 fi
 
-rm "${__sub_file}"
+if [ -e "${__sub_file}" ]; then
+    rm "${__sub_file}"
+fi
 
 exit
